+++ Office365/main.py
@@ -2,6 +2,7 @@ from sekoia_automation.module import Module
 
 from office365.management_api.connector import Office365Connector
 from office365.message_trace.trigger_office365_messagetrace import Office365MessageTraceTrigger
+from office365.message_trace.trigger_office365_messagetrace_graph_api import Office365MessageTraceTriggerGraphAPI
 from office365.message_trace.trigger_office365_messagetrace_oauth import (
     Office365MessageTraceTrigger as Office365MessageTraceTriggerOAuth,
 )
@@ -11,4 +12,5 @@ if __name__ == "__main__":
     module.register(Office365Connector, "office365_management_api")
     module.register(Office365MessageTraceTrigger, "office365_messagetrace_trigger")
     module.register(Office365MessageTraceTriggerOAuth, "office365_messagetrace_trigger_oauth")
+    module.register(Office365MessageTraceTriggerGraphAPI, "office365_messagetrace_trigger_graph_api")
     module.run()