+++ Office365/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-03-05 - 2.20.0
+
+### Added
+
+- Add Office 365 Message Trace Graph API connector
+
+### Changed
+
+- Upgrade dependencies
+- Upgrade sekoia-automation-sdk
+
 ## 2025-12-19 - 2.19.1
 
 ### Changed